back_end/file_service/requirements.txt
Django==4.2.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
gunicorn==21.2.0
celery==5.3.6
redis==5.0.1